Conflicts:
	eTrade_ejb/src/main/java/fr/dauphine/etrade/model/Actualite.java
	eTrade_ejb/src/main/java/fr/dauphine/etrade/model/Societe.java
	eTrade_ejb/src/main/java/fr/dauphine/etrade/services/ServicesUtilisateurBean.java

+++ eTrade_ejb/src/main/java/fr/dauphine/etrade/api/ServicesUtilisateur.java
@@ -12,7 +12,7 @@ public interface ServicesUtilisateur {
 	Utilisateur addUtilisateur(Utilisateur utilisateur);
 	Utilisateur delUtilisateur(Utilisateur utilisateur);
 	List<Utilisateur> allUtilisateurs();
-	Utilisateur getUtilisateurById(int id);
+	Utilisateur getUtilisateurById(long id);
 	Utilisateur getUtilisateurByEmail(String email);
 	Utilisateur updateUtilisateur(Utilisateur utilisateur);
 	List<Utilisateur> getUnvalidatedUtilisateurs();
